Renaming feedback for data lake, first iteration.
This renames all of the data lake commands to dla and dls
It breaks them into separate modules
It updates our params to more nicely follow the '-' standard.
Updates tests to use new naming.
Updates help to use new naming.
