--- src/command_modules/azure-cli-datalake/HISTORY.rst
@@ -1,14 +1,0 @@
-.. :changelog:
-
-Release History
-===============
-
-0.0.1b2 (2017-03-13)
-+++++++++++++++++++++
-
-* Fix: Add datalake package to setup.py (#2493)
-
-0.0.1b1 (2017-03-13)
-+++++++++++++++++++++
-
-* Initial Preview Release for Azure Data Lake Store and Analytics commands
