--- src/command_modules/azure-cli-datalake/README.rst
@@ -1,7 +1,0 @@
-Microsoft Azure CLI 'datalake' Command Module
-=============================================
-
-This package is for the 'datalake' module.
-i.e. 'az datalake'
-
-
