--- src/command_modules/azure-cli-datalake/azure/cli/command_modules/datalake/analytics/__init__.py
@@ -1,7 +1,0 @@
-# --------------------------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License. See License.txt in the project root for license information.
-# --------------------------------------------------------------------------------------------
-def load_params(_):
-    #pylint: disable=redefined-outer-name
-    import azure.cli.command_modules.datalake.analytics._params
