+++ src/command_modules/azure-cli-dla/MANIFEST.in
