+++ src/command_modules/azure-cli-dla/README.rst
@@ -1,0 +1,7 @@
+Microsoft Azure CLI 'data lake analytics' Command Module
+=============================================
+
+This package is for the 'data lake analytics' module.
+i.e. 'az dla'
+
+
