+++ src/command_modules/azure-cli-dla/tests/recordings/test_dla_catalog_mgmt.yaml
