+++ src/command_modules/azure-cli-dls/HISTORY.rst
@@ -1,0 +1,18 @@
+.. :changelog:
+
+Release History
+===============
+0.0.1b3 (2017-03-13)
++++++++++++++++++++++
+
+* Rename and separate command modules for Data Lake Analytics and Data Lake Store
+
+0.0.1b2 (2017-03-13)
++++++++++++++++++++++
+
+* Fix: Add datalake package to setup.py (#2493)
+
+0.0.1b1 (2017-03-13)
++++++++++++++++++++++
+
+* Initial Preview Release for Azure Data Lake Store and Analytics commands
