+++ src/command_modules/azure-cli-dls/MANIFEST.in
@@ -1,0 +1,1 @@
+include *.rst
