+++ src/command_modules/azure-cli-dls/README.rst
@@ -1,0 +1,7 @@
+Microsoft Azure CLI 'data lake store' Command Module
+=============================================
+
+This package is for the 'data lake store' module.
+i.e. 'az dls'
+
+
